Crate brakets

Strongly-typed bra-ket notation for discrete quantum objects.

Structs

Bra
Generic Bra. You can multiply it by ket to get an inner product (scalar).

Ket
Generic ket. You can multiply it by bra to an outer product (or operator).

Outer
Outer product in bra-ket notation, as well used as a linear operatior. You can multiply it by ket to put it in the observable state defined by the operator.

U8
A type level dimension.

U16
A type level dimension.

U32
A type level dimension.

U64
A type level dimension.

Type Definitions

Bra2
2-dimension bra

Bra4
4-dimension bra (for entangled states)

Complex

Ket2
2-dimension ket

Ket4
4-dimension ket (for entangled states)

Outer2
2-dimension outer

Outer4
4-dimension ket (for entangled states)